gadgethub/serializers.py
from rest_framework import serializers
from django.contrib.auth.models import User
from rest_framework_simplejwt.tokens import RefreshToken
from .models import Product


class UserSerializer(serializers.ModelSerializer):
    name = serializers.SerializerMethodField(read_only=True)
    isSeller = serializers.SerializerMethodField(read_only=True)
    
    class Meta:
        model = User
        fields = ['id','username','email','name','isSeller']
        
    def get_name(self,obj):
        name = obj.first_name
        
        if name == "":
            name = obj.email
        
        return name
    
    def get_isSeller(self,obj):
        
        return obj.is_staff
    
    
class UserWithRefreshToken(UserSerializer):
    token = serializers.SerializerMethodField(read_only=True)
    
    class Meta:
        model = User
        fields = ['id','username','email','name','isSeller','token']
          
    def get_token(self,obj):
        token = RefreshToken.for_user(obj)
        return str(token.access_token)
        
class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = '__all__'